feat: swap joins the app shell; visible entries on You + Settings
/swap moves into the (app) route group (URL unchanged) so it inherits
AppShell, PageHeader and the mm system like every sibling: mm-panel
figure rows, mm-field amount + mm-btn MAX, wire staged actions, kicker
stepper, semantic result states. swap.css rewritten onto --mm-* tokens;
legacy sheets and .sc-root dropped, quarantine test back to grant-only.

Entries: You Money & control gets a primary Swap ETH -> USDG button
(first, beside Wallet & permissions), and the Settings auto-convert
section gets a prominent entry button showing on/off state. The tiny
checkbox row stays as the toggle.
